loginapp/requirements.txt
fastapi
starlette>=0.27.0
uvicorn
jinja2
pydantic
python-multipart
python-dotenv
qrcodegen
Pillow
asyncpg
aiofiles